third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFETurbulenceElement-svgdom-baseFrequency-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'baseFrequency' property of the SVGFETurbulenceElment object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Math.round(turbulence.baseFrequencyX.baseVal * 1000) is 70
PASS Math.round(turbulence.baseFrequencyY.baseVal * 1000) is 70
PASS Math.round(turbulence.baseFrequencyX.baseVal * 1000) is 50
PASS Math.round(turbulence.baseFrequencyY.baseVal * 1000) is 50
PASS successfullyParsed is true

TEST COMPLETE
